blurg/requirements.txt
django-tagging
markdown
